moving experience functions to a separate file. Also added the ability to load the the experience questions on page load
